Update ElectronChannel_plots.py and MuonChannel_plots.py: Add lumis. Add DoubleMuon_2016F to Samples_cern.dat. Suppress 0 xmin value in Canvas.py. Fix overflow bin in Samples.py (int to float in bin width). Add yshift parameter in MakeFullPlot function in Utils.py.

+++ include/Sample.py
@@ -138,7 +138,7 @@ class Sample:
           h_of = TH1F(name, "", _nbins+1, _newarr)                                     
       else:
            h = TH1F(name+'_noOF', "", nbin, xmin, xmax)
-           bw = int((xmax-xmin)/nbin)
+           bw = float((xmax-xmin)/nbin)
            ylabel = ylabel +"/ " + str(bw) + " GeV"
            h_of = TH1F(name, '', nbin+1, xmin, xmax+bw)
       h.Sumw2()
